Stopped tracking dotenv file and added template dotenv file. Allowed for specifying delete paramter for clearing messages.

+++ src/app.js
@@ -54,7 +54,9 @@ async function handleOnMessage(msg) {
 
     try {
         await commandExecutor.execute(msg, args);
-        await msg.delete();
+        if (!commandExecutor.persistCommand) {
+            await msg.delete();
+        }
     } catch (error) {
         console.error(error);
         msg.reply(msgConsts.EXECUTION_ERROR);
